Use token instead of email

+++ app/actions.ts
@@ -3,10 +3,16 @@
 import client from "@/lib/mongodb";
 import { ObjectId } from "mongodb";
 import { Todo } from "../app/interfaces";
+import { decodeJWT } from "./util";
 
 const collection = client.db("todo").collection<Todo>("todo");
 
-export const getTasks = async (email : string) : Promise<Todo[]> => {
+const authorise = (token : string) => {
+  return decodeJWT(token).email;
+};
+
+export const getTasks = async (token : string) : Promise<Todo[]> => {
+  const email = authorise(token);
   try {
     const data = await collection
       .find({ owner: email })
@@ -20,7 +26,8 @@ export const getTasks = async (email : string) : Promise<Todo[]> => {
   }
 };
 
-export const addTask = async (email : string, newTaskName : string) => {
+export const addTask = async (token : string, newTaskName : string) => {
+  const email = authorise(token);
     const newTodo : Todo = {
         _id: new ObjectId(),
         name: newTaskName,
@@ -33,12 +40,14 @@ export const addTask = async (email : string, newTaskName : string) => {
     return await getTasks(email);
 };
 
-export const toggleTask = async (id: string, done: boolean) => {
+export const toggleTask = async (token: string, id: string, done: boolean) => {
   console.log(`Marking todo ${id} as ${done}`);
-  const data = await collection.updateOne({ _id: new ObjectId(id) }, { $set: { done } });
+  const email = authorise(token);
+  const data = await collection.updateOne({ _id: new ObjectId(id), owner: email }, { $set: { done } });
   return JSON.parse(JSON.stringify(data));
 };
 
-export const deleteTask = async (id: string) => {
-  await collection.deleteOne({ _id: new ObjectId(id) });
+export const deleteTask = async (token : string, id: string) => {
+  const email = authorise(token);
+  await collection.deleteOne({ _id: new ObjectId(id), owner: email });
 };
